Rename Matrix.py to matrix.py
